Make view spec less picky about attribute order
The order in which Rails' `link_to` helper method orders attributes can differ
from setup to setup. This can cause a failing spec, depending on which computer
is running the specs. Fix by making the spec look within the link tag for the
specific attributes we care about: href, class, id, and link text.

This approach is shown in RSpec's documentation:

https://www.relishapp.com/rspec/rspec-rails/docs/helper-specs/helper-spec