Fix strategy_spec to stub call to Janrain refresh_token API
The strategy_spec fails with this error:

  1) Devise::Capturable if user does not exist and
     capturable_redirect_if_no_user is enabled should redirect

     Failure/Error: @strategy.authenticate!
     NoMethodError:
       undefined method `capturable_server' for Devise:Module
     # ./lib/devise_capturable/api.rb:22:in `refresh_token'
     # ./lib/devise_capturable/strategy.rb:46:in `authenticate!'
     # ./spec/strategy_spec.rb:79:in `block (4 levels) in <top (required)>'

The problem is not that capturable_server is unconfigured, but that the
API is unstubbed.  Commit b46365df added a call to the refresh_token
API, which and is where this test case starts to fail.

Stub the API call and expect the new redirect target (with refresh
token).