Make view helper test more resilient to environmental differences
The spec for the link_to_capturable view helper appears to be dependent
on some part of the Ruby development environment. On my setup and
Marcel's, the spec passes when we compare output against

    <a href="#" class="capture_modal_open" id="capture_signin_link">Login</a>

However, in Rune's environment, the passing output is

    <a class="capture_modal_open" href="#" id="capture_signin_link">Login</a>

To address this discrepancy, break down the test to look for matches
within the string, regardless of the order in which they get rendered.