Ignore test logs and database file
The embedded test Rails application outputs to a log file and stores
test data in a .sqlite3 file. While these files would typically be
ignored in a standard Rails application's Git repository, they are not
ignored in this testing structure. They may be accidentally committed
into source control as a result.

Ignore log/*.log and db/*.sqlite3 files in Git.