Package Capture-related functionality into Capturable extension
We'd like to make Capturable the go-to solution for Janrain logins in
O'Reilly's Rails applications. However, it is currently broken (does
not properly complete a Janrain login), and requires application
developers to replicate boilerplate code to render sign-in and sign-out
widgets.

Package as much of the Capture-related functionality as possible into
the Capturable extension, including:

* A built-in view to render the sign-in widget, with the required
  JavaScript
* A sessions controller that expects the Janrain authorization code by
  Ajax, and returns the access token back to the browser to properly
  complete Janrain login and resolve the "refresh" issue we've seen in
  apps using the existing Capturable.
* A helper to show/hide the sign-out widget depending on the page being
  viewed
* Support for Janrain's password reset functionality
* Support for Federated sign-out
* Support for O'Reilly's fix for partial JavaScript blockers

https://oreillymedia.atlassian.net/browse/ACONF-252