Improve support for logout and modal widgets
The Janrain widget must be present on a page in order for logout to
work. While the current implementation works on the customer support
portal, or any application requiring login system-wide, it will not
work on apps that have public-facing pages. Update the logic to not
display the logout widget when a user is not logged in.

As a side benefit, this now makes implementation of the modal login
widget relatively trivial. Reconfigure the partials to include the
modal widget, and alter the Ajax functionality to only redirect when a
login request comes from the embedded widget.

One known issue with this implementation: The logout link's display is
currently dependent on the login state in Rails, not in JavaScript, so
the link doesn't redirect upon successful modal login.

https://oreillymedia.atlassian.net/browse/ACONF-1497