Make sure @resource_name is accessible in JavaScript
We rely on the the Devise resource's name, set to @resource_name, to
properly point to server-side login in a Capturable-enabled app. This
value is currently not available and causing test suites to fail.

Fix by making sure @resource_name gets defined where needed. Extract
the somewhat obscure lookup of Devise mappings to a helper.