Add handling for application-side failure messages
Sometimes, the application using Capturable will have a non-Janrain
error that requires special messaging, and is not fixable by a page
refresh. Add client-side support for this messaging.

Note: Providing the messaging (by adding failure_message to the results
returned to the client) currently requires overriding the sessions
controller in the application.