Allow error message popup to be closed
If an error occurs during sign-in, the user is presented with a modal
error message. The only way to clear this message is to reload the
page, and other links on the page are no longer accessible. Add a link
to close this message and restore page functionality. Fix a heretofore
unnoticed bug in the message display function.

https://oreillymedia.atlassian.net/browse/ACONF-2257