Only push data to Umbel if _umbel is defined
If Backplane is enabled, but Umbel is not configured on the page on
which the Backplane event is triggered, JavaScript will throw an error
(_umbel is not defined). This caused a lot of unexpected broken tests
in an application using Capturable, since Umbel support is added
dynamically there.

Add a conditional to check whether _umbel is defined inside the event
handler. If it is, then proceed. If not, then provide info to
console.log that Backplane triggered an event, but was not able to do
anything with Umbel. Move the Backplane logic to a partial, which can
be included anywhere in Capturable or a host application.