Use updating widget loaders provided by Janrain
Janrain has changed how the sign-in widget is loaded into the browser.
Instead of providing a hard-coded cloudfront.net URL, we can now use
Janrain's "standard" widget loaders. This makes sure any future changes
made to the sign-in widget in the Janrain dashboard (for example, adding
or editing social login providers) will be reflected in applications
using the widget. Instead of systematically building the loader URL
based on protocol, Janrain now suggests providing a full URL for both
HTTPS and HTTP-based versions of the widget.

Replace the original JavaScript for loading the correct version of the
widget with updated code, provided by Janrain. Update the README to
specify how to apply the two URLs to an application using Capturable.