Remove references to Backplane and Umbel in Noscript messaging
O'Reilly is no longer contracting with Umbel, so we don't need
Backplane or Umbel as part of an optimal user experience across sites.
Remove the suggestion to allow related domains from the messaging we
show JavaScript blockers like NoScript and Ghostery.

The remaining Backplane-related code does not affect anything if
Backplane is not enabled on a given site using Janrain, so leave it in
place to maintain backward compatibility.

https://oreillymedia.atlassian.net/browse/ACONF-6023
https://oreillymedia.atlassian.net/browse/ACONF-6015