Load Janrain login widget JavaScript per-locale
We want the text in the Janrain login widget to be translated, and some
text is hardcoded in the JavaScript, it doesn't come from Janrain.

Internationalize the path to janrain_login_widget, so we can point to
translations of the JavaScript. Add new files on the CDN at the correct
locations. This will pair with a change on Expectnation's dev branch to
point to the staged version of the translated login widget. Update the
README to describe the new functionality this gives us to override the
janrain_widget_url per-locale as well as per-environment.

https://oreillymedia.atlassian.net/browse/ACONF-6368
https://oreillymedia.atlassian.net/browse/ACONF-6261