Extract customer service contact email to allow override
We're not sure which contact email will be used for Chinese events, and
it would be nice if we didn't need to change devise-capturable as well
as the host application to change this email address.

Extract email into its own key, so that it can be overridden from the
host app.

https://oreillymedia.atlassian.net/browse/ACONF-6261
https://oreillymedia.atlassian.net/browse/ACONF-6275